requirements.txt
torch
open-clip-torch
faiss-cpu
Pillow
httpx==0.24.1
fastapi==0.103.2
uvicorn[standard]==0.30.6
python-multipart
pydantic==1.10.17
starlette==0.27.0
openai
python-dotenv
